+++ simpleble/src/backends/android/BackendAndroid.h
@@ -5,6 +5,8 @@
 
 namespace SimpleBLE {
 
+class AdapterAndroid;
+
 class BackendAndroid : public BackendSingleton<BackendAndroid> {
   public:
     BackendAndroid(buildToken);
@@ -15,6 +17,10 @@ class BackendAndroid : public BackendSingleton<BackendAndroid> {
     virtual std::string name() const noexcept override;
 
   private:
+    // Android devices only have a single Bluetooth adapter, so in order to preserve
+    // state across multiple instances, a single Adapter object is shared across
+    // all users of this backend.
+    std::shared_ptr<AdapterAndroid> _adapter;
 };
 
 }  // namespace SimpleBLE