+++ docs/changelog.rst
@@ -20,6 +20,7 @@ The format is based on `Keep a Changelog`_, and this project adheres to `Semanti
 **Changed**
 
 - (MacOS) Use a single Adapter object across all users of the CoreBluetooth backend.
+- (Android) Use a single Adapter object across all users of the Android backend.
 
 **Fixed**